figma-transformer: extract instance override resolver
